+++ main.py
@@ -3,12 +3,12 @@ import streamlit as st
 from sibylapp2.compute.context import get_term
 from sibylapp2.config import get_pages_to_show
 from sibylapp2.pages import (
-    Compare_Entities,
-    Experiment_with_Changes,
+    # Compare_Entities,
+    # Experiment_with_Changes,
     Explore_a_Prediction,
-    Settings,
-    Similar_Entities,
-    Understand_the_Model,
+    # Settings,
+    # Similar_Entities,
+    # Understand_the_Model,
 )
 from sibylapp2.view.utils import setup
 
@@ -16,11 +16,11 @@ setup.setup_page(return_row_ids=True)
 
 ALL_PAGES = {
     "Explore a Prediction": Explore_a_Prediction,
-    "Similar Entities": Similar_Entities,
-    "Compare Entities": Compare_Entities,
-    "Experiment with Changes": Experiment_with_Changes,
-    "Understand the Model": Understand_the_Model,
-    "Settings": Settings,
+    # "Similar Entities": Similar_Entities,
+    # "Compare Entities": Compare_Entities,
+    # "Experiment with Changes": Experiment_with_Changes,
+    # "Understand the Model": Understand_the_Model,
+    # "Settings": Settings,
 }
 
 pages_to_show = get_pages_to_show()
