+++ sibylapp2/config.yml
@@ -17,5 +17,6 @@ NEGATIVE_TERM: # Term to use for a False prediction for boolean models
 PRED_FORMAT_STRING: # fstring for formatting numeric model outputs
 OVERRIDE_PRED_FORMAT: # If True, use "config.manual_pred_format_func" to format preds, ignoring previous 3 configs
 SUPPORT_PROBABILITY: # Only set this to True when model supports `predict_proba`
-PAGES_TO_SHOW: # Set to a list of pages or "all" to show all pages
+PAGES_TO_SHOW:
+- Explore a Prediction # Set to a list of pages or "all" to show all pages
 ALLOW_PAGE_SELECTION:  # If true, allow user to select which pages to show (requires settings page)
