+++ sibylapp2/pages/Compare_Entities.py
@@ -3,8 +3,8 @@
 import streamlit as st
 
 from sibylapp2.compute.context import get_term
-from sibylapp2.view import compare_entities
-from sibylapp2.view.utils import display, filtering
+from sibylapp2.view import compare_entities, filtering
+from sibylapp2.view.utils import display
 
 
 def main():
