+++ sibylapp2/pages/Settings.py
@@ -8,7 +8,7 @@ from ruamel import yaml
 from sibylapp2 import config
 from sibylapp2.compute.contributions import get_dataset_contributions
 from sibylapp2.compute.model import get_dataset_predictions
-from sibylapp2.view.utils.helpers import neg_em, pos_em
+from sibylapp2.view.helpers import neg_em, pos_em
 
 UP_ARROW = "⬆"
 DOWN_ARROW = "⬇"
