+++ sibylapp2/view/compare_entities.py
@@ -4,7 +4,7 @@ import streamlit as st
 from sibylapp2.compute import contributions, features, model
 from sibylapp2.compute.context import get_term
 from sibylapp2.config import POSITIVE_TERM, PREDICTION_TYPE, PredType, pred_format_func
-from sibylapp2.view.utils import filtering, helpers
+from sibylapp2.view import filtering, helpers
 from sibylapp2.view.utils.formatting import format_two_contributions_to_view
