+++ sibylapp2/view/global_contributions.py
@@ -5,8 +5,8 @@ from pyreal.visualize import strip_plot
 
 from sibylapp2.compute import contributions, features
 from sibylapp2.compute.context import get_term
-from sibylapp2.view.utils import filtering, helpers
-from sibylapp2.view.utils.helpers import show_legend
+from sibylapp2.view import filtering, helpers
+from sibylapp2.view.helpers import show_legend
 
 
 @st.cache_data(show_spinner="Generating summary plot...")
