--- sibylapp2/view/utils/display.py
@@ -1,35 +1,0 @@
-import streamlit as st
-
-from sibylapp2.compute import context, model
-from sibylapp2.config import SUPPORT_PROBABILITY, pred_format_func
-
-
-def show_probability_select_box():
-    if SUPPORT_PROBABILITY:
-        st.sidebar.checkbox(
-            "Show probability",
-            value=st.session_state["display_proba"],
-            help="Display prediction in terms of probability",
-            key="display_proba",
-        )
-
-
-def view_prediction(eid):
-    predictions = model.get_predictions(
-        st.session_state["eids"], model_id=st.session_state["model_id"]
-    )
-    if SUPPORT_PROBABILITY:
-        predictions_proba = model.get_predictions(
-            st.session_state["eids"], model_id=st.session_state["model_id"], return_proba=True
-        )
-
-    pred = predictions[eid]
-    if st.session_state["display_proba"]:
-        pred_proba = predictions_proba[eid]
-        pred_display = (
-            f"{pred_format_func(pred)} ({pred_format_func(pred_proba, display_proba=True)})"
-        )
-    else:
-        pred_display = pred_format_func(pred)
-
-    st.sidebar.metric(context.get_term("Prediction"), pred_display)
